% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beamer_marco.R
\name{beamer_metropolis}
\alias{beamer_metropolis}
\title{Metropolis (beamer) presentation with sensible defaults.}
\usage{
beamer_metropolis(slide_level = 2, toc = FALSE, incremental = FALSE, ...)
}
\arguments{
\item{slide_level}{By default 2}

\item{toc}{By default FALSE}

\item{incremental}{By default FALSE}

\item{...}{Other params passed to \link[bookdown:html_document2]{bookdown::beamer_presentation2}}
}
\value{
An R Markdown output format object to be passed to \code{\link[rmarkdown:render]{rmarkdown::render()}}.
}
\description{
Wrapper for \link[bookdown:html_document2]{bookdown::beamer_presentation2} with sensible defaults.
}
