% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bib_subset}
\alias{bib_subset}
\title{Subsets a bib(la)tex file with the bibliography actually used in an Rmd file}
\usage{
bib_subset(rmdfile, bibfile, newfile = NULL)
}
\arguments{
\item{rmdfile}{An Rmd file.}

\item{bibfile}{Its bib(la)tex file.}

\item{newfile}{The new bib(la)tex file.}
}
\value{
The new bibliography in a file if specified
}
\description{
Subsets a bib(la)tex file with the bibliography actually used in an Rmd file. Adapted from https://stackoverflow.com/questions/65306015/in-rmarkdown-is-there-a-way-to-create-a-bib-file-for-only-those-keys-cited-in
}
