% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emoji.R
\name{emoji}
\alias{emoji}
\title{Add an emoji in an Rmarkdown document.}
\usage{
emoji(x, folder = "emoji")
}
\arguments{
\item{x}{Either name or keyword. If more than one emoji has the specified keyword, will pick one at random.}

\item{folder}{Folder to store the emojis. Only relevant in a Rmd file compiled in latex.}
}
\description{
Add an emoji to an Rmarkdown file. This is a wrapper of \link[emo:ji]{emo::ji} that will download the emoji and will add it as an image, if the Rmarkdown document will be compiled in latex to pdf.  Code adapted from Hadley Wickham's code in \link{https://github.com/hadley/adv-r/blob/master/emoji.R}. Use \link[emo:ji_name]{emo::ji_name} to get a list of unique names.
}
\examples{
emoji("face")
 

}
